Alias args are now parsed and saved as a list.
This should've been done in the first place.
Arguments from cmdline are already in a list,
so just pass that to cmd_alias_add() and let
it save them as a list (using JSON).

This also makes it easier to include file names
or arguments with spaces. Just make sure you
are saving valid JSON in cedit.conf.